check authorities in crud requests

+++ src/main/java/db/api/rest_controllers/CrudAthleteController.java
@@ -3,6 +3,7 @@ package db.api.rest_controllers;
 import db.api.service.AthleteService;
 import db.entities.Athlete;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -13,19 +14,28 @@ public class CrudAthleteController {
 
     @PostMapping("")
     public Athlete addAthlete(@RequestParam("firstName") String firstName, @RequestParam("patronymic") String patronymic,
-                              @RequestParam("lastName") String lastName, @RequestParam("club") String title) {
+                              @RequestParam("lastName") String lastName, @RequestParam("club") String title, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return athleteService.addAthlete(firstName, patronymic, lastName, title);
     }
 
     @PutMapping("")
     public Athlete updateAthlete(@RequestParam("id") Long id, @RequestParam("firstName") String firstName,
                                  @RequestParam("patronymic") String patronymic, @RequestParam("lastName") String lastName,
-                                 @RequestParam("club") String title) {
+                                 @RequestParam("club") String title, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return athleteService.updateAthlete(id, firstName, patronymic, lastName, title);
     }
 
     @DeleteMapping("")
-    public Athlete deleteAthlete(@RequestParam("athlete") Long athleteId) {
+    public Athlete deleteAthlete(@RequestParam("athlete") Long athleteId, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return athleteService.deleteAthlete(athleteId);
     }
 }
