+++ src/main/java/db/api/rest_controllers/CrudAthleteRankingController.java
@@ -3,6 +3,7 @@ package db.api.rest_controllers;
 import db.api.service.AthleteRankingService;
 import db.entities.AthleteRanking;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -13,19 +14,28 @@ public class CrudAthleteRankingController {
 
     @PostMapping("")
     public AthleteRanking addAthleteRankAndSport(@RequestParam("id") Long athleteId, @RequestParam("rank") String rank,
-                                                 @RequestParam("sport") String value) {
+                                                 @RequestParam("sport") String value, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return athleteRankingService.addAthleteRanking(athleteId, rank, value);
     }
 
     @PutMapping("")
     public AthleteRanking updateAthleteRankAndSport(@RequestParam("id") Long athleteId, @RequestParam("rank") String rank,
-                                                 @RequestParam("sport") Long sportId) {
+                                                 @RequestParam("sport") Long sportId, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return athleteRankingService.updateAthleteRanking(athleteId, sportId, rank);
     }
 
     @DeleteMapping("")
     public AthleteRanking deleteAthleteRankAndSport(@RequestParam("athlete") Long athleteId,
-                                                 @RequestParam("sport") Long sportId) {
+                                                 @RequestParam("sport") Long sportId, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return athleteRankingService.deleteAthleteRanking(athleteId, sportId);
     }
 }
