+++ src/main/java/db/api/rest_controllers/CrudCompetitionPlayerController.java
@@ -3,6 +3,7 @@ package db.api.rest_controllers;
 import db.api.service.CompetitionPlayerService;
 import db.entities.CompetitionPlayer;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -15,7 +16,10 @@ public class CrudCompetitionPlayerController {
     public CompetitionPlayer addCompetitionPlayer(@RequestParam("athlete") Long athleteId,
                                                   @RequestParam("competition") Long competitionId,
                                                   @RequestParam("wasawarding") boolean wasawarding,
-                                                  @RequestParam("result") Long result) {
+                                                  @RequestParam("result") Long result, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return competitionPlayerService.addCompetitionPlayer(athleteId, competitionId, wasawarding, result);
     }
 
@@ -23,13 +27,19 @@ public class CrudCompetitionPlayerController {
     public CompetitionPlayer updateCompetitionPlayer(@RequestParam("athlete") Long athleteId,
                                                   @RequestParam("competition") Long competitionId,
                                                   @RequestParam("wasawarding") boolean wasawarding,
-                                                  @RequestParam("result") Long result) {
+                                                  @RequestParam("result") Long result, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return competitionPlayerService.updateCompetitionPlayer(athleteId, competitionId, wasawarding, result);
     }
 
     @DeleteMapping("")
     public CompetitionPlayer deleteCompetitionPlayer(@RequestParam("athlete") Long athleteId,
-                                                     @RequestParam("competition") Long competitionId) {
+                                                     @RequestParam("competition") Long competitionId, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return competitionPlayerService.deleteCompetitionPlayer(athleteId, competitionId);
     }
 }
