+++ src/main/java/db/api/rest_controllers/CrudSportController.java
@@ -3,6 +3,7 @@ package db.api.rest_controllers;
 import db.api.service.SportService;
 import db.entities.Sport;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -12,17 +13,27 @@ public class CrudSportController {
     private final SportService sportService;
 
     @PostMapping("")
-    public Sport addSport(@RequestParam("sport") String value) {
+    public Sport addSport(@RequestParam("sport") String value, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return sportService.addSport(value);
     }
 
     @DeleteMapping("")
-    public Sport deleteSport(@RequestParam("sport") Long sportId) {
+    public Sport deleteSport(@RequestParam("sport") Long sportId, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return sportService.deleteSport(sportId);
     }
 
     @PutMapping("")
-    public Sport updateSport(@RequestParam("id") Long id, @RequestParam("sport") String sport) {
+    public Sport updateSport(@RequestParam("id") Long id,
+                             @RequestParam("sport") String sport, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return sportService.updateSport(id, sport);
     }
 }
