+++ src/main/java/db/api/rest_controllers/CrudTrainerController.java
@@ -3,6 +3,7 @@ package db.api.rest_controllers;
 import db.api.service.TrainerService;
 import db.entities.Trainer;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -13,18 +14,28 @@ public class CrudTrainerController {
 
     @PostMapping("")
     public Trainer addTrainer(@RequestParam("firstName") String firstName, @RequestParam("patronymic") String patronymic,
-                              @RequestParam("lastName") String lastName) {
+                              @RequestParam("lastName") String lastName, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return trainerService.addTrainer(firstName, patronymic, lastName);
     }
 
     @PutMapping("")
     public Trainer updateTrainer(@RequestParam("id") Long id, @RequestParam("firstName") String firstName,
-                                 @RequestParam("patronymic") String patronymic, @RequestParam("lastName") String lastName) {
+                                 @RequestParam("patronymic") String patronymic, @RequestParam("lastName") String lastName,
+                                 Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return trainerService.updateTrainer(id, firstName, patronymic, lastName);
     }
 
     @DeleteMapping("")
-    public Trainer deleteTrainer(@RequestParam("trainer") Long trainerId) {
+    public Trainer deleteTrainer(@RequestParam("trainer") Long trainerId, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return trainerService.deleteTrainer(trainerId);
     }
 }
