+++ src/main/java/db/api/rest_controllers/CrudUserController.java
@@ -3,6 +3,8 @@ package db.api.rest_controllers;
 import db.api.service.user.CustomUserDetailsServiceImpl;
 import db.entities.user.User;
 import lombok.AllArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -12,7 +14,10 @@ public class CrudUserController {
     private final CustomUserDetailsServiceImpl customUserDetailsService;
 
     @DeleteMapping("")
-    public User deleteUser(@RequestParam("user") Long userId) {
+    public User deleteUser(@RequestParam("user") Long userId, Authentication authentication) {
+        if (authentication.getAuthorities().stream().noneMatch(a -> a.getAuthority().equals("ADMIN"))) {
+            return null;
+        }
         return customUserDetailsService.deleteUser(userId);
     }
